Add SPI for ShardingSphereYAMLTupleProcessor
