+++ shardingsphere-underlying/shardingsphere-common/src/main/java/org/apache/shardingsphere/underlying/common/log/ConfigurationLogger.java
@@ -20,15 +20,13 @@ package org.apache.shardingsphere.underlying.common.log;
 import lombok.AccessLevel;
 import lombok.NoArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.apache.shardingsphere.underlying.common.yaml.representer.processor.NoneTupleProcessor;
-import org.apache.shardingsphere.underlying.common.yaml.swapper.RuleRootConfigurationsYamlSwapper;
 import org.apache.shardingsphere.underlying.common.auth.Authentication;
 import org.apache.shardingsphere.underlying.common.auth.yaml.swapper.AuthenticationYamlSwapper;
 import org.apache.shardingsphere.underlying.common.config.RuleConfiguration;
 import org.apache.shardingsphere.underlying.common.yaml.engine.YamlEngine;
+import org.apache.shardingsphere.underlying.common.yaml.swapper.RuleRootConfigurationsYamlSwapper;
 
 import java.util.Collection;
-import java.util.Collections;
 import java.util.Properties;
 
 /**
@@ -44,7 +42,7 @@ public final class ConfigurationLogger {
      * @param ruleConfigurations rule configurations
      */
     public static void log(final Collection<RuleConfiguration> ruleConfigurations) {
-        log("Rule configurations: ", YamlEngine.marshal(new RuleRootConfigurationsYamlSwapper().swap(ruleConfigurations), Collections.singleton(new NoneTupleProcessor())));
+        log("Rule configurations: ", YamlEngine.marshal(new RuleRootConfigurationsYamlSwapper().swap(ruleConfigurations)));
     }
     
     /**
